requirements.txt
PyMuPDF==1.23.22
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
opencv-python==4.11.0.86
scikit-learn>=1.5
threadpoolctl>=3.5